backend/requirements.txt
# Core LangChain dependencies
langchain-core>=0.3.0
langchain>=0.3.0
langgraph>=0.2.0

# LLM providers
langchain-anthropic>=0.2.0
langchain-openai>=0.2.0
langchain-aws>=0.1.0

# Additional utilities
python-dotenv>=1.0.0
boto3>=1.34.0

# Web scraping for search tool
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0